apps/vendor-analysis/pyproject.toml
[project]
name = "vendor-analysis"
version = "0.1.0"
description = "Read-only NetSuite vendor cost analysis and management CLI"
requires-python = ">=3.12"
dependencies = [
    "libs-logger @ file:///opt/ns/libs/logger",
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pandas>=2.2.0",
    "rich>=13.7.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pyright>=1.1.0",
]
# Note: Rich is in main dependencies (used by CLI + bootstrap.py)
# Bootstrap script runs BEFORE uv sync, so Rich must be installed separately first

[project.scripts]
vendor-analysis = "vendor_analysis.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "SIM", "TCH"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
reportMissingTypeStubs = true
reportUnknownMemberType = true
reportUnknownArgumentType = true
reportUnknownVariableType = true
strictListInference = true
strictDictionaryInference = true
strictSetInference = true

